requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
gradio>=4.0.0
pillow>=10.0.0
imageio>=2.31.0
matplotlib>=3.7.0
torch==1.13.1
torchvision==0.14.1
scikit-image==0.21.0
opencv-contrib-python>=4.8.0
pygame>=2.5.0